agents/bennutritionniste.ai/requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic
ffmpeg-python
pydub
tqdm
pillow
pytesseract
whisper
python-dotenv
jinja2
chromadb
python-docx
instagrapi
google-generativeai
google-cloud-aiplatform
google-cloud-storage